# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Cloudexus API",
        "version": "1",
        "summary": "The catalog, stock, prices, invoices, partners and orders, and the warehouse app's bookings.",
        "description": "JSON in and out; every message in English. Amounts are in the primary currency (meta.currency); translatable text comes in the language asked for (meta.language). The whole description, for people, is web/API.md — and inside the application under API → API documentation.",
        "license": {
            "name": "AGPL-3.0-only",
            "identifier": "AGPL-3.0-only"
        }
    },
    "servers": [
        {
            "url": "https://cloudexus.example/api"
        }
    ],
    "security": [
        {
            "token": []
        },
        {
            "apiKey": []
        }
    ],
...
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "operationId": "products",
                "summary": "Products",
                "description": "q searches SKU, name and barcode.",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    },
                    {
                        "$ref": "#/components/parameters/q"
                    },
                    {
                        "$ref": "#/components/parameters/updated_since"
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "One category."
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        },
                        "description": "Active or inactive only."
                    },
                    {
                        "$ref": "#/components/parameters/language"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A page.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PageMeta"
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/Unprocessable"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
...
        "/products/{id}": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "operationId": "product",
                "summary": "A product",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    },
                    {
                        "$ref": "#/components/parameters/language"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The product.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductDetail"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Meta"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "scheme": "bearer",
                "description": "An integration token (API → API users) or a user token from POST /auth/login (cxu_…)."
            },
            "apiKey": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Api-Key",
                "description": "The same token, for servers that strip the Authorization header."
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                },
                "description": "The record's id."
            },
            "page": {
                "name": "page",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 200,
                    "default": 50
                }
            },
            "q": {
                "name": "q",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string"
                },
                "description": "Free-text search; the fields searched differ per list."
            },
            "updated_since": {
                "name": "updated_since",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string",
                    "examples": [
                        "2026-07-19 00:00:00"
                    ]
                },
                "description": "Only records changed since then: for a delta sync."
            },
            "language": {
                "name": "language",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "string",
                    "examples": [
                        "en"
                    ]
                },
                "description": "The language translatable text comes in (see GET /languages); a missing translation falls back to the default. An unknown code is 422."
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "Invalid or missing API token.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
            "NotFound": {
                "description": "Not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
            "Unprocessable": {
                "description": "The request does not make sense; details may say which lines.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "TooManyRequests": {
                "description": "Rate limit exceeded: 60 requests a minute per token (per user for user tokens).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "object",
                        "properties": {
                            "status": {
                                "type": "integer",
                                "description": "The HTTP status again."
                            },
                            "message": {
                                "type": "string",
                                "description": "In English."
                            },
                            "details": {
                                "description": "For a client to act on: which lines are short of stock, two_factor_required."
                            }
                        },
                        "required": [
                            "status",
                            "message"
                        ]
                    }
                },
                "required": [
                    "error"
                ]
            },
            "Currency": {
                "description": "The primary currency every amount is in. Amounts are never converted.",
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "examples": [
                            "HUF"
                        ]
                    },
                    "symbol": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "symbol",
                    "title"
                ]
            },
            "Language": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "examples": [
                            "en"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "default": {
                        "type": "string",
                        "description": "The default language's code, which a missing translation falls back to."
                    }
                },
                "required": [
                    "code",
                    "name",
                    "default"
                ]
            },
            "Meta": {
                "type": "object",
                "properties": {
                    "currency": {
                        "$ref": "#/components/schemas/Currency"
                    },
                    "language": {
                        "$ref": "#/components/schemas/Language"
                    }
                }
            },
            "PageMeta": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    },
                    "currency": {
                        "$ref": "#/components/schemas/Currency"
                    },
                    "language": {
                        "$ref": "#/components/schemas/Language"
                    }
                },
                "required": [
                    "page",
                    "per_page",
                    "total",
                    "total_pages"
                ]
            },
...
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "barcode": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "description": "In the language asked for."
                    },
                    "short_description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "HTML."
                    },
                    "category_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "category_name": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "In lists."
                    },
                    "unit_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "unit": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The unit's code."
                    },
                    "unit_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "price": {
                        "type": "string",
                        "description": "Net.",
                        "pattern": "^-?[0-9]+(\\.[0-9]+)?$",
                        "examples": [
                            "89900.00"
                        ]
                    },
                    "sale_price": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Net; when set, the active price.",
                        "pattern": "^-?[0-9]+(\\.[0-9]+)?$",
                        "examples": [
                            "89900.00"
                        ]
                    },
                    "vat_rate": {
                        "type": "string",
                        "pattern": "^-?[0-9]+(\\.[0-9]+)?$",
                        "examples": [
                            "89900.00"
                        ]
                    },
                    "min_stock": {
                        "type": "string",
                        "pattern": "^-?[0-9]+(\\.[0-9]+)?$",
                        "examples": [
                            "89900.00"
                        ]
                    },
                    "is_active": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ]
                    },
                    "is_webshop": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ]
                    },
                    "width_mm": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "height_mm": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "depth_mm": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "weight_g": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "stock_qty": {
                        "type": [
                            "number",
                            "string"
                        ],
                        "description": "Current stock, all warehouses."
                    },
                    "created_at": {
                        "type": "string",
                        "examples": [
                            "2026-07-24 16:44:35"
                        ]
                    },
                    "updated_at": {
                        "type": "string",
                        "examples": [
                            "2026-07-24 16:44:35"
                        ]
                    }
                },
                "required": [
                    "id",
                    "sku",
                    "name",
                    "price",
                    "is_active"
                ]
            },
            "ProductDetail": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Product"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "images": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "path": {
                                            "type": "string",
                                            "description": "A relative upload path or a full URL."
                                        },
                                        "is_primary": {
                                            "type": "integer",
                                            "enum": [
                                                0,
                                                1
                                            ]
                                        },
                                        "sort_order": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            },
                            "attributes": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "parameter_id": {
                                            "type": "integer"
                                        },
                                        "attr_name": {
                                            "type": "string"
                                        },
                                        "attr_value": {
                                            "type": "string"
                                        },
                                        "sort_order": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            },
                            "category_ids": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            },
                            "related_ids": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            },
                            "substitute_ids": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                }
                            },
                            "group_prices": {
                                "type": "object",
                                "description": "By customer_group_id.",
                                "additionalProperties": {
                                    "type": "object",
                                    "properties": {
                                        "customer_group_id": {
                                            "type": "integer"
                                        },
                                        "price": {
                                            "type": "string",
                                            "pattern": "^-?[0-9]+(\\.[0-9]+)?$",
                                            "examples": [
                                                "89900.00"
                                            ]
                                        },
                                        "sale_price": {
                                            "type": [
                                                "string",
                                                "null"
                                            ],
                                            "pattern": "^-?[0-9]+(\\.[0-9]+)?$",
                                            "examples": [
                                                "89900.00"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    }
                ]
            },
...
        }
    }
}